feat: Refactor E2E tests for Android and iOS; consolidate checks and improve structure
