+++ .github/workflows/checks.yml
@@ -66,8 +66,8 @@ jobs:
 
       - name: Configure kubectl
         run: |
-          echo "${{ secrets.KUBECONFIG_FOR_E2E_TESTS_BASE64 }}" | base64 -d > ci-e2e-reader.kubeconfig
-          echo "KUBECONFIG=$GITHUB_WORKSPACE/ci-e2e-reader.kubeconfig" >> $GITHUB_ENV
+          echo "${{ secrets.KUBECONFIG_FOR_E2E_TESTS_BASE64 }}" | base64 -d > apps/yral-mobile/ci-e2e-reader.kubeconfig
+          echo "KUBECONFIG=$GITHUB_WORKSPACE/apps/yral-mobile/ci-e2e-reader.kubeconfig" >> $GITHUB_ENV
 
       - name: Fetch Kafka password
         run: |
@@ -77,41 +77,26 @@ jobs:
           echo "KAFKA_PASSWORD=$PASS" >> $GITHUB_ENV
 
       # ══════════════════════════════════════════════════════════════════════
-      # ALL CHECKS  (lint · unit tests · iOS xcodebuild · Android APK build)
+      # ALL CHECKS
+      # The emulator runner creates the AVD and starts it before the script
+      # runs. Checks.androidStartEmulator() detects it already running and
+      # skips the start step. Everything else — lint, unit tests, iOS
+      # xcodebuild, and both Android + iOS e2e — runs in a single invocation.
       # ══════════════════════════════════════════════════════════════════════
 
       - name: Run all checks
-        working-directory: apps/yral-mobile
-        run: ./gradlew allChecks
-
-      - name: Upload detekt SARIF
-        if: always()
-        uses: github/codeql-action/upload-sarif@v4
-        with:
-          sarif_file: apps/yral-mobile/build/reports/detekt/merge.sarif
-
-      # ══════════════════════════════════════════════════════════════════════
-      # E2E TESTS
-      # ══════════════════════════════════════════════════════════════════════
-
-      - name: Android E2E — feed scroll + Kafka assert
         uses: reactivecircus/android-emulator-runner@v2
         with:
           api-level: 34
           arch: x86_64
           script: |
             cd apps/yral-mobile
-            E2E_PLATFORM=android \
-            E2E_APP_ID=com.yral.android \
             KAFKA_BOOTSTRAP=kafka.yral.com:9092 \
             KAFKA_PASSWORD=$KAFKA_PASSWORD \
-            ./gradlew :checks:test
+            ./gradlew allChecks
 
-      - name: iOS E2E — feed scroll + Kafka assert
-        working-directory: apps/yral-mobile
-        run: |
-          E2E_PLATFORM=ios \
-          E2E_APP_ID=com.yral.iosApp.staging \
-          KAFKA_BOOTSTRAP=kafka.yral.com:9092 \
-          KAFKA_PASSWORD=$KAFKA_PASSWORD \
-          ./gradlew :checks:test
+      - name: Upload detekt SARIF
+        if: always()
+        uses: github/codeql-action/upload-sarif@v4
+        with:
+          sarif_file: apps/yral-mobile/build/reports/detekt/merge.sarif
